Updata paypal 5-12

+++ THDShop/THDShop/DE_ORDER.cs
@@ -23,7 +23,7 @@ namespace THDShop
         public int IDORDER { get; set; }
         public int IDPRODUCT { get; set; }
         public int QUANTITY { get; set; }
-        public double PRICE { get; set; }
+        public int PRICE { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<DE_BILL> DE_BILL { get; set; }
